Expanded RSR IPC for BMR intergration improvements, update logic for AutoAttack and CancelCast logic
Split auto-attack and cast-cancel handling into new AutoAttackUpdater and CancelCastUpdater classes for better modularity. Removed related logic from ActionQueueManager and MiscUpdater. Updated initialization, disposal, and main update loops to use the new updaters. Cleaned up code and improved logging.
