+++ RotationSolver.Basic/Helpers/TargetFilter.cs
@@ -233,7 +233,7 @@ public static class TargetFilter
 		foreach (T obj in objects)
 		{
 			if (obj.DistanceToPlayer() <= radius)
-					yield return obj;
+				yield return obj;
 		}
 	}
 }
