+++ RotationSolver/Updaters/ActionQueueManager.cs
@@ -3,8 +3,6 @@ using ECommons.DalamudServices;
 using ECommons.GameHelpers;
 using ECommons.Logging;
 using FFXIVClientStructs.FFXIV.Client.Game;
-using FFXIVClientStructs.FFXIV.Client.Game.UI;
-using RotationSolver.Basic.Configuration;
 using RotationSolver.Commands;
 
 namespace RotationSolver.Updaters
@@ -14,12 +12,8 @@ namespace RotationSolver.Updaters
 		// Action Manager Hook for intercepting user input
 		private static Hook<UseActionDelegate>? _useActionHook;
 
-		// Hook for AutoAttackState.SetImpl to prevent auto attacks
-		private static Hook<SetAutoAttackStateDelegate>? _setAutoAttackStateHook;
-
-		// Delegates for ActionManager functions
+		// Delegate for ActionManager UseAction
 		private unsafe delegate bool UseActionDelegate(ActionManager* actionManager, uint actionType, uint actionID, ulong targetObjectID, uint param, uint useType, int pvp, bool* isGroundTarget);
-		private unsafe delegate bool SetAutoAttackStateDelegate(AutoAttackState* self, bool value, bool sendPacket, bool isInstant);
 
 		public static void Enable()
 		{
@@ -81,10 +75,6 @@ namespace RotationSolver.Updaters
 				_useActionHook = Svc.Hook.HookFromAddress<UseActionDelegate>(useActionAddress, UseActionDetour);
 				_useActionHook?.Enable();
 
-				var setAutoAttackStateAddress = AutoAttackState.Addresses.SetImpl.Value;
-				_setAutoAttackStateHook = Svc.Hook.HookFromAddress<SetAutoAttackStateDelegate>(setAutoAttackStateAddress, SetAutoAttackStateDetour);
-				_setAutoAttackStateHook?.Enable();
-
 				PluginLog.Debug("[ActionQueueManager] Action interception hooks initialized");
 			}
 			catch (Exception ex)
@@ -101,86 +91,12 @@ namespace RotationSolver.Updaters
 				_useActionHook?.Dispose();
 				_useActionHook = null;
 
-				_setAutoAttackStateHook?.Disable();
-				_setAutoAttackStateHook?.Dispose();
-				_setAutoAttackStateHook = null;
-
 				PluginLog.Debug("[ActionQueueManager] Action interception hooks disposed");
 			}
 			catch (Exception ex)
 			{
 				PluginLog.Error($"[ActionQueueManager] Failed to dispose action hooks: {ex}");
 			}
-		}
-
-		/// <summary>
-		/// Called every frame. If auto attacks are currently active but a NoCastingStatus is
-		/// present, sends the toggle-auto-attack general action to disable them.
-		/// </summary>
-		public static unsafe void Update()
-		{
-			if (!Player.Available) return;
-
-			try
-			{
-				var uiState = UIState.Instance();
-				if (uiState == null) return;
-
-				if (uiState->WeaponState.AutoAttackState.IsAutoAttacking && PlayerHasNoCastingStatus())
-				{
-					// GeneralAction 1 is the auto-attack toggle — same method the game uses
-					ActionManager.Instance()->UseAction(ActionType.GeneralAction, 1);
-					PluginLog.Information("[ActionQueueManager] Disabled active auto attacks due to NoCastingStatus.");
-				}
-			}
-			catch (Exception ex)
-			{
-				PluginLog.Warning($"[ActionQueueManager] Error in Update (auto attack disable): {ex.Message}");
-			}
-		}
-
-		private static bool PlayerHasNoCastingStatus()
-		{
-			try
-			{
-				HashSet<uint> noCastingStatus = OtherConfiguration.NoCastingStatus;
-				if (noCastingStatus == null || noCastingStatus.Count == 0) return false;
-				if (Player.Object?.StatusList == null) return false;
-
-				// Check for Motion Tracker seperately, idk its weird
-				if (StatusHelper.PlayerHasStatus(false, StatusID.MotionTracker))
-				{
-					return true;
-				}
-
-				if (DataCenter.BMRSpecialModeType == SpecialMode.Pyretic)
-				{
-					return true;
-				}
-
-				foreach (var status in Player.Object.StatusList)
-				{
-					if (noCastingStatus.Contains(status.StatusId))
-						return true;
-				}
-			}
-			catch (Exception ex)
-			{
-				PluginLog.Warning($"[ActionQueueManager] Error checking NoCastingStatus: {ex.Message}");
-			}
-			return false;
-		}
-
-		private static unsafe bool SetAutoAttackStateDetour(AutoAttackState* self, bool value, bool sendPacket, bool isInstant)
-		{
-			// Block attempts to enable auto attacks while a NoCastingStatus is active
-			if (value && Player.Available && PlayerHasNoCastingStatus())
-			{
-				PluginLog.Debug("[ActionQueueManager] Prevented auto attack activation due to NoCastingStatus.");
-				return true;
-			}
-
-			return _setAutoAttackStateHook!.Original(self, value, sendPacket, isInstant);
 		}
 
 		private static unsafe bool UseActionDetour(ActionManager* actionManager, uint actionType, uint actionID, ulong targetObjectID, uint param, uint useType, int pvp, bool* isGroundTarget)
