+++ RotationSolver/Updaters/MiscUpdater.cs
@@ -4,22 +4,19 @@ using Dalamud.Game.Text.SeStringHandling.Payloads;
 using ECommons.DalamudServices;
 using ECommons.ExcelServices;
 using ECommons.GameHelpers;
-using FFXIVClientStructs.FFXIV.Client.Game.UI;
 using FFXIVClientStructs.FFXIV.Client.System.Framework;
 using FFXIVClientStructs.FFXIV.Client.UI;
 using FFXIVClientStructs.FFXIV.Client.UI.Misc;
-using RotationSolver.Basic.Configuration;
 using RotationSolver.Commands;
 
 namespace RotationSolver.Updaters;
 
 internal static class MiscUpdater
 {
-
 	internal static void UpdateMisc()
 	{
 		UpdateEntry();
-		UpdateCancelCast();
+		CancelCastUpdater.UpdateCancelCast();
 	}
 
 	private static IDtrBarEntry? _dtrEntry;
@@ -132,94 +129,7 @@ internal static class MiscUpdater
 		};
 	}
 
-	private static RandomDelay _tarStopCastDelay = new(() => Service.Config.StopCastingDelay);
-
-	private static unsafe void UpdateCancelCast()
-	{
-		if (Player.Object == null || !Player.Object.IsCasting)
-		{
-			return;
-		}
-
-		if (!DataCenter.State)
-		{
-			return;
-		}
-
-		IBattleChara? castTarget = Svc.Objects.SearchById(Player.Object.CastTargetObjectId) as IBattleChara;
-
-		bool tarDead = Service.Config.UseStopCasting
-			&& castTarget != null
-			&& castTarget.IsEnemy()
-			&& castTarget.CurrentHp == 0;
-
-		// Cancel raise cast if target already has Raise status
-		bool tarHasRaise = castTarget != null && castTarget.HasStatus(false, StatusID.Raise);
-
-		// Cancel cast in PvP if the target gains Guard and the action does not ignore Guard
-		bool tarHasGuard = DataCenter.IsPvP && Service.Config.PvpGuardCancel
-			&& castTarget != null
-			&& castTarget.HasStatus(false, StatusID.Guard)
-			&& !(((ActionID)Player.Object.CastActionId).GetActionFromID(true, RotationUpdater.CurrentRotationActions)
-				is IBaseAction guardCheckAction && (!guardCheckAction.Setting.IgnoreGuard || (DataCenter.Job == Job.BLM && !guardCheckAction.Setting.IgnoreGuard && !StatusHelper.PlayerHasStatus(true, StatusID.WreathOfFire))));
-
-		float[] statusTimes = GetStatusTimes();
-
-		float minStatusTime = float.MaxValue;
-		for (int i = 0; i < statusTimes.Length; i++)
-		{
-			if (statusTimes[i] < minStatusTime)
-			{
-				minStatusTime = statusTimes[i];
-			}
-		}
-
-		float remainingCast = MathF.Max(0, Player.Object.TotalCastTime - Player.Object.CurrentCastTime);
-
-		// Cancel immediately if the player currently has any active NoCastingStatus
-		bool hasNoCastingStatus = statusTimes.Length > 0;
-
-		// Cancel if a "no-casting" status will expire before the cast completes and it's soon (<3s)
-		bool stopDueStatus = hasNoCastingStatus
-			&& minStatusTime <= remainingCast
-			&& minStatusTime < 3f;
-
-		bool shouldStopHealing =
-			Service.Config.StopHealingAfterThresholdExperimental2
-			&& DataCenter.InCombat
-			&& !CustomRotation.HealingWhileDoingNothing
-			&& DataCenter.CommandNextAction?.AdjustedID != Player.Object.CastActionId
-			&& ((ActionID)Player.Object.CastActionId).GetActionFromID(true, RotationUpdater.CurrentRotationActions)
-				is IBaseAction { Setting.GCDSingleHeal: true }
-			&& (DataCenter.MergedStatus & (AutoStatus.HealAreaSpell | AutoStatus.HealSingleSpell)) == 0;
-
-		if (_tarStopCastDelay.Delay(tarDead) || hasNoCastingStatus || stopDueStatus || tarHasRaise || tarHasGuard || shouldStopHealing)
-		{
-			UIState* uiState = UIState.Instance();
-			if (uiState != null)
-			{
-				uiState->Hotbar.CancelCast();
-			}
-		}
-	}
-
-	private static float[] GetStatusTimes()
-	{
-		List<float> statusTimes = [];
-		if (Player.Object?.StatusList != null)
-		{
-			foreach (Dalamud.Game.ClientState.Statuses.IStatus status in Player.Object.StatusList)
-			{
-				if (OtherConfiguration.NoCastingStatus.Contains(status.StatusId))
-				{
-					statusTimes.Add(status.RemainingTime);
-				}
-			}
-		}
-		return [.. statusTimes];
-	}
-
-	internal static unsafe void PulseActionBar(uint actionID)
+	internal static void PulseActionBar(uint actionID)
 	{
 		LoopAllSlotBar((bar, hot, index) =>
 		{
@@ -227,7 +137,7 @@ internal static class MiscUpdater
 		});
 	}
 
-	private static unsafe bool IsActionSlotRight(ActionBarSlot slot, RaptureHotbarModule.HotbarSlot? hot, uint actionID)
+	private static bool IsActionSlotRight(ActionBarSlot slot, RaptureHotbarModule.HotbarSlot? hot, uint actionID)
 	{
 		if (hot.HasValue)
 		{
@@ -255,7 +165,7 @@ internal static class MiscUpdater
 		return Service.GetAdjustedActionId((uint)slot.ActionId) == actionID;
 	}
 
-	private unsafe delegate bool ActionBarAction(ActionBarSlot bar, RaptureHotbarModule.HotbarSlot? hot, uint highLightID);
+	private delegate bool ActionBarAction(ActionBarSlot bar, RaptureHotbarModule.HotbarSlot? hot, uint highLightID);
 	private unsafe delegate bool ActionBarPredicate(ActionBarSlot bar, RaptureHotbarModule.HotbarSlot* hot);
 	private static unsafe void LoopAllSlotBar(ActionBarAction doingSomething)
 	{
@@ -309,7 +219,7 @@ internal static class MiscUpdater
 		}
 	}
 
-	public static unsafe void Dispose()
+	public static void Dispose()
 	{
 		if (_dtrEntry?.Title != null)
 		{
